remove: frontend from project (#13)
removes the frontend from the repo and does all the necessary
refactoring to keep things alive.
