--- frontend/postcss.config.js
@@ -1,7 +1,0 @@
-// postcss.config.js (ESM)
-import tailwindcss from '@tailwindcss/postcss'
-import autoprefixer from 'autoprefixer'
-
-export default {
-  plugins: [tailwindcss(), autoprefixer()],
-}
