--- frontend/src/index.css
@@ -1,5 +1,0 @@
-/* src/index.css */
-@import "tailwindcss";
-
-/* optional helper */
-.json-pre { white-space: pre-wrap; word-break: break-word; font-size: 0.9rem; line-height: 1.5; }
