+++ requirements.txt
@@ -1,5 +1,3 @@
--e .
-
 Flask
 Flask-Cors
 Pillow
